Vessel manager, Client should have dropdown suggestions from previous or saved entries
Vessel Vessel Active / Discontinued Info and check box and Discontinue remarks to be added with status history showing same. Refer to HazinventM

Not able to save a new vessel, "discontinueRemarks" is not allowed error coming.

Not able to update a vessel details, "discontinueRemarks" is not allowed error coming.

Docker image optimized for smaller build.